+++ example/lib/post_frame_callback_example/view_model/my_view_model_with_post_frame_callback.dart
@@ -4,28 +4,26 @@ import 'package:view_model_x/view_model_x.dart';
 class MyViewModelWithPostFrameCallback extends ViewModel
     with PostFrameCallback {
   // initialize SharedFlow
-  final _messageSharedFlow = MutableSharedFlow<String>();
-
-  SharedFlow<String> get messageSharedFlow => _messageSharedFlow;
+  final messageSharedFlow = SharedFlow<String>();
 
   @override
   void init() {
     // do stuff on create of view model, (equivalent to constructor)
     debugPrint("on init");
 
-    _messageSharedFlow.emit(
+    messageSharedFlow.emit(
         "on init"); // this emit will not received by listener because it emitted before ui build.
   }
 
   @override
   void onPostFrameCallback(Duration timestamp) {
-    _messageSharedFlow.emit(
+    messageSharedFlow.emit(
         "onPostFrameCallback"); // this emit will be received because it is emitted after the ui build completed.
   }
 
   @override
   void dispose() {
-    _messageSharedFlow.dispose();
+    messageSharedFlow.dispose();
     debugPrint("MyViewModelWithPostFrameCallback disposed");
   }
 }