+++ example/lib/responsive_example/view_model/second_view_model.dart
@@ -3,14 +3,12 @@ import 'package:view_model_x/view_model_x.dart';
 
 class SecondViewModel extends ViewModel {
   // initialize SharedFlow
-  final _messageSharedFlow = MutableSharedFlow<String>();
-
-  SharedFlow<String> get messageSharedFlow => _messageSharedFlow;
+  final messageSharedFlow = SharedFlow<String>();
 
   void showPopupMessage() {
     // by emitting the value, listeners were notified
     debugPrint("hi");
-    _messageSharedFlow.emit("Hello from MyViewModel!");
+    messageSharedFlow.emit("Hello from MyViewModel!");
   }
 
   @override
@@ -20,7 +18,7 @@ class SecondViewModel extends ViewModel {
 
   @override
   void dispose() {
-    _messageSharedFlow.dispose();
+    messageSharedFlow.dispose();
     debugPrint("SecondViewModel disposed");
   }
 }